+++ lib/capybara_accessible_selectors/selectors/role.rb
@@ -4,7 +4,7 @@ require "capybara_accessible_selectors/aria"
 
 Capybara.add_selector :role do
   xpath do |locator, custom_elements: nil, **|
-    roles = expand_roles(locator)
+    roles = CapybaraAccessibleSelectors::Helpers.expand_roles(locator)
     [
       *roles.flat_map { CapybaraAccessibleSelectors::Aria::IMPLICIT_ROLE_SELECTORS[_1.to_sym] || [] },
       *Array(custom_elements).map { XPath.descendant(_1.to_sym) }.map { _1[!XPath.attr(:role)] },
@@ -13,15 +13,7 @@ Capybara.add_selector :role do
   end
 
   locator_filter do |node, locator, **|
-    expand_roles(locator).include?(node.role)
-  end
-
-  def expand_roles(locator)
-    roles = Array(locator).map(&:to_s)
-    CapybaraAccessibleSelectors::Aria::ROLE_SYNONYMS.each do |s|
-      roles.push(*s) if s.intersect?(roles)
-    end
-    roles.uniq
+    CapybaraAccessibleSelectors::Helpers.expand_roles(locator).include?(node.role)
   end
 
   filter_set(:capybara_accessible_selectors, %i[aria described_by])