+++ spec/matchers/have_image_spec.rb
@@ -1,16 +1,12 @@
 # frozen_string_literal: true
 
-describe "have_img" do
-  before do
-    allow(Capybara::Helpers).to receive(:warn)
-  end
-
+describe "have_image" do
   it "matches using a custom matcher" do
     render <<~HTML
       <img alt="Img title" src="data:image/gif;base64,R0lGODlhAQABAAAAACH5BAEKAAEALAAAAAABAAEAAAICTAEAOw==" width="10" height="10" data-test-id="test">
     HTML
 
-    expect(page).to have_img "Img title"
+    expect(page).to have_image "Img title"
   end
 
   it "matches using a negated custom matcher" do
@@ -18,6 +14,6 @@ describe "have_img" do
       <img alt="Img title" src="data:image/gif;base64,R0lGODlhAQABAAAAACH5BAEKAAEALAAAAAABAAEAAAICTAEAOw==" width="10" height="10" data-test-id="test">
     HTML
 
-    expect(page).to have_no_img "foo"
+    expect(page).to have_no_image "foo"
   end
 end